Add files via upload
Barbieri recreation files - notebook and character-based embedding data processing. 
BLSTM4 is a recreation of Barbieri's Bidirecitonal LSTM in the paper "Are Emojis Predictable?" that we are using as a baseline. We used the same dataset of tweets as Barbieri, but was unable to determine the character-based encoding he used.  We created a pre-trained character-based encoding by further processing GloVe, which is a word-based embedding. The processing is done in the 'create_char_embedded_dataset.py' file which is called by the main BLSTM4 notebook.